requirements.txt
python-telegram-bot==21.6
requests==2.31.0
python-dotenv==1.0.0

# Testing
pytest==9.0.2
pytest-asyncio==0.23.0
pytest-cov==6.0.0

# Profiling & Performance (install when needed)
# memory-profiler==0.61.0
# line-profiler==4.1.0
